// src/app/_models/models-enum.ts
export enum deleted_status {
  Yes = "Yes",
  No = "No"
}

export enum user_status {
  Active = "Active",
  InActive = "In Active"
}
export enum image_type {
  Image = "image",
  Video = "video"
}
export enum student_status {
  Pending = "Pending",
  Approved = "Approved",
  Rejected = "Rejected"
}